requirements.txt
Flask==2.2.2
gunicorn==20.1.0
requests==2.28.1
python-dotenv==0.21.0
openai==0.25.0
pytest==7.2.0